feat: ScoringHighlightsMatch class
