+++ .gitignore
@@ -23,5 +23,6 @@
 hs_err_pid*
 
 # Others
+.idea/
 target/
 *.iml
