+++ pom.xml
@@ -2,34 +2,79 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0"
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
+  <modelVersion>4.0.0</modelVersion>
 
-    <groupId>lt.jocas</groupId>
-    <artifactId>lucene-monitor-helpers</artifactId>
-    <version>1.0-SNAPSHOT</version>
+  <groupId>lt.jocas</groupId>
+  <artifactId>lucene-monitor-helpers</artifactId>
+  <version>0.1.0-SNAPSHOT</version>
+  <name>lucene-monitor-helpers</name>
 
-    <properties>
-        <maven.compiler.source>11</maven.compiler.source>
-        <maven.compiler.target>11</maven.compiler.target>
-    </properties>
+  <description>Missing classes of the Lucene Monitor</description>
+  <url>https://github.com/dainiusjocas/clj-jq</url>
 
-    <dependencies>
-        <dependency>
-            <groupId>org.apache.lucene</groupId>
-            <artifactId>lucene-core</artifactId>
-            <version>8.7.0</version>
-        </dependency>
-        <dependency>
-            <groupId>org.apache.lucene</groupId>
-            <artifactId>lucene-monitor</artifactId>
-            <version>8.7.0</version>
-        </dependency>
-        <dependency>
-            <groupId>com.oracle.substratevm</groupId>
-            <artifactId>svm</artifactId>
-            <version>19.2.1</version>
-            <scope>provided</scope>
-        </dependency>
-    </dependencies>
+  <licenses>
+    <license>
+      <name>The Apache License, Version 2.0</name>
+      <url>http://www.apache.org/licenses/LICENSE-2.0.txt</url>
+    </license>
+  </licenses>
+
+  <properties>
+    <maven.compiler.source>11</maven.compiler.source>
+    <maven.compiler.target>11</maven.compiler.target>
+  </properties>
+
+  <developers>
+    <developer>
+      <name>Dainius Jocas</name>
+      <email>dainius.jocas@gmail.com</email>
+      <organization>jocas.lt</organization>
+      <organizationUrl>http://www.jocas.lt</organizationUrl>
+    </developer>
+  </developers>
+
+  <dependencies>
+    <dependency>
+      <groupId>org.apache.lucene</groupId>
+      <artifactId>lucene-core</artifactId>
+      <version>8.8.1</version>
+    </dependency>
+    <dependency>
+      <groupId>org.apache.lucene</groupId>
+      <artifactId>lucene-monitor</artifactId>
+      <version>8.8.1</version>
+    </dependency>
+  </dependencies>
+
+  <build>
+    <sourceDirectory>src</sourceDirectory>
+    <resources>
+      <resource>
+        <directory>src</directory>
+      </resource>
+    </resources>
+  </build>
+
+  <scm>
+    <connection>scm:git:git://github.com/dainiusjocas/lucene-monitor-helpers.git</connection>
+    <developerConnection>scm:git:ssh://git@github.com/dainiusjocas/lucene-monitor-helpers.git</developerConnection>
+    <tag>main</tag>
+    <url>https://github.com/dainiusjocas/lucene-monitor-helpers</url>
+  </scm>
+
+  <repositories>
+    <repository>
+      <id>clojars</id>
+      <url>https://repo.clojars.org/</url>
+    </repository>
+  </repositories>
+
+  <distributionManagement>
+    <repository>
+      <id>clojars</id>
+      <name>Clojars repository</name>
+      <url>https://clojars.org/repo</url>
+    </repository>
+  </distributionManagement>
 
 </project>
